Move most visualization logic into viz app
