+++ gateway/config/api_router.py
@@ -7,8 +7,8 @@ from sds_gateway.api_methods.views.capture_endpoints import CaptureViewSet
 from sds_gateway.api_methods.views.dataset_endpoints import DatasetViewSet
 from sds_gateway.api_methods.views.file_endpoints import FileViewSet
 from sds_gateway.api_methods.views.file_endpoints import check_contents_exist
-from sds_gateway.api_methods.views.visualization_endpoints import VisualizationViewSet
 from sds_gateway.users.api.views import UserViewSet
+from sds_gateway.visualizations.api_views import VisualizationViewSet
 
 router = DefaultRouter() if settings.DEBUG else SimpleRouter()
