+++ gateway/sds_gateway/visualizations/management/__init__.py
@@ -1,0 +1,1 @@
+# Management package for visualizations app
