+++ gateway/sds_gateway/visualizations/management/commands/__init__.py
@@ -1,0 +1,1 @@
+# Commands package for visualizations app
