+++ gateway/sds_gateway/visualizations/management/commands/setup_pipelines.py
@@ -1,4 +1,4 @@
-"""Management command to set up django-cog pipelines."""
+"""Management command to set up django-cog visualization pipelines."""
 
 import datetime
 import json
@@ -10,21 +10,20 @@ from django_cog.models import Pipeline
 from django_cog.models import Stage
 from django_cog.models import Task
 
-from sds_gateway.api_methods.cog_pipelines import PIPELINE_CONFIGS
-from sds_gateway.api_methods.models import get_latest_pipeline_by_base_name
+from sds_gateway.visualizations.cog_pipelines import PIPELINE_CONFIGS
 
 
 class Command(BaseCommand):
-    """Set up django-cog pipelines for post-processing."""
+    """Set up django-cog pipelines for visualization processing."""
 
-    help = "Set up django-cog pipelines for post-processing"
+    help = "Set up django-cog pipelines for visualization processing"
 
     def add_arguments(self, parser):
         """Add command arguments."""
         parser.add_argument(
             "--pipeline-type",
             type=str,
-            choices=["waterfall", "all"],
+            choices=["visualization", "all"],
             default="all",
             help="Type of pipeline to set up",
         )
@@ -175,6 +174,8 @@ class Command(BaseCommand):
         config = config_func()
 
         # Check if pipeline already exists (including versioned pipelines)
+        from sds_gateway.visualizations.models import get_latest_pipeline_by_base_name
+
         existing_pipeline = get_latest_pipeline_by_base_name(config["pipeline_name"])
         if existing_pipeline:
             pipeline = existing_pipeline
