+++ gateway/sds_gateway/visualizations/processing/__init__.py
@@ -1,0 +1,9 @@
+"""Processing subpackage for visualizations."""
+
+from .spectrogram import generate_spectrogram_from_drf
+from .waterfall import convert_drf_to_waterfall_json
+
+__all__ = [
+    "convert_drf_to_waterfall_json",
+    "generate_spectrogram_from_drf",
+]
